poking aroung

+++ common/volBucket/bcluster.c
@@ -6,7 +6,7 @@
   #include "volBucket.h"
 #include "time.h"
 
-  /* 
+  /*
 
   C = bcluster(X, r);
 
@@ -23,7 +23,7 @@
 
   except from the output format.
 
-  Can also be run standalone, for debugging, 
+  Can also be run standalone, for debugging,
   gcc bcluster.c volBucket.c
   ./a.out
 
@@ -32,7 +32,7 @@ microscopy image stack) into a regular grid of buckets. Each point is
 assigned to bucket and is linked to buckets that might contain other
 points with d<r.
 
-delta, the side length of the buckets, has to be larger than rmax, 
+delta, the side length of the buckets, has to be larger than rmax,
 or some neighbours might not be
 found
 
@@ -57,13 +57,13 @@ void propagate(vbBucket * myVB, vbNode *e, uint32 * C, uint32 * cpos)
     while(b!= NULL)
     {
       if(b->visited == 0)
-       if(vbBucketDistance2(b,e) < myVB->rr)     
+       if(vbBucketDistance2(b,e) < myVB->rr)
          propagate(myVB, b, C, cpos);
-      b=b->next;  
+      b=b->next;
     }
   }
-}    
- 
+}
+
 
 void bcluster(double *X, int nObjects, double r, uint32 * C)
 /* Input: X, the data [[x0, y0, z0], [x1, y1, z1], ... ,[xN, yN, zN]]
@@ -74,11 +74,11 @@ void bcluster(double *X, int nObjects, double r, uint32 * C)
 int verbose = 0; // 0: none, 1: some, 2: insane
 // Get from X
 
-int maxx=ceil(X[0]); 
+int maxx=ceil(X[0]);
 int maxy=ceil(X[nObjects]);
 int maxz=ceil(X[2*nObjects]);
 
-int minx=floor(X[0]); 
+int minx=floor(X[0]);
 int miny=floor(X[nObjects]);
 int minz=floor(X[2*nObjects]);
 
@@ -114,12 +114,12 @@ if(verbose)
 
 for(uint32 kk=0; kk<nObjects; kk++)
 {
-  double x = X[kk]; 
-  double y = X[kk+nObjects]; 
-  double z = X[kk+2*nObjects]; 
+  double x = X[kk];
+  double y = X[kk+nObjects];
+  double z = X[kk+2*nObjects];
   if(verbose>1)
    printf("%f, %f, %f\n", x, y, z);
-  vbAdd(myVB, x,y,z, NULL, kk+1); 
+  vbAdd(myVB, x,y,z, NULL, kk+1);
   if(verbose>1)
     printf("%u\n", myVB->nObjects);
 }
@@ -132,7 +132,7 @@ for(uint32 kk=0; kk<nObjects; kk++) // For each node/object
   {
    vbNode *e = &myVB->objects[kk];
    // printf("p:%d v:%d enr: %d\n", cpos, e->visited, e->number);
-   int cpos0 = cpos; 
+   int cpos0 = cpos;
    if (e->visited == 0)
       {
         propagate(myVB, e, C, &cpos);
@@ -140,7 +140,7 @@ for(uint32 kk=0; kk<nObjects; kk++) // For each node/object
       }
 
   // Exclude non - clusters
-  if(cpos == cpos0+2) 
+  if(cpos == cpos0+2)
     {
        C[cpos0]=0;
        cpos = cpos0;
@@ -158,7 +158,7 @@ int main(int argv, char ** argc)
 printf("sizeof(uint32): %lu\n", sizeof(uint32));
 
 srand(time(NULL)); rand();
-int N = 10000+ (double) (5000.0*rand()/RAND_MAX);
+int N = 10000 + (double) (5000.0*rand()/RAND_MAX);
 N = 1;
 double r = 1.0 + ((double) rand()/(RAND_MAX));
 
@@ -170,9 +170,9 @@ memset(C, 0, 2*N*sizeof(uint32));
 
 for(int kk=0; kk<N; kk++)
 {
-  X[kk] = 1024.0*rand()/RAND_MAX;
-  X[kk+N] = 1024.0*rand()/RAND_MAX;
-  X[kk+2*N] = 20.0*rand()/RAND_MAX;
+    X[kk] = 1024.0*rand()/ (double) RAND_MAX;
+    X[kk+N] = 1024.0*rand()/ (double) RAND_MAX;
+    X[kk+2*N] = 20.0*rand()/ (double) RAND_MAX;
 }
 X[0] = 20; X[1]=1; X[2] = 1;
 
@@ -184,7 +184,7 @@ for(uint32 kk=0; kk<2*N; kk++)
   {
   if(C[kk] == 0)
       { if(last>0) printf("\n"); }
-      else 
+      else
       {printf("%u ", C[kk]);}
   last = C[kk];
 }
@@ -198,4 +198,3 @@ free(C);
 
 return 0;
 }
-
